Hungnm

+++ onlinelearning/src/main/java/group5/swp391/onlinelearning/controller/teacher/CourseController.java
@@ -65,16 +65,23 @@ public class CourseController {
     public String getCourseList(Model model, HttpSession req) {
         // TODO: remove user service
         User user = userService.getUserById(70);
-        req.setAttribute("studentSession", user);
+        req.setAttribute("userSession", user);
+        // Check role access site
+        if (user.getRole() != 1)
+            return "AccessDenied";
+
         List<CourseDTOTeacher> courses = courseService.getCourseDTOTeacherList();
         String title = "Course List";
         thymeleafBaseCRUD.setBaseForList(model, courses, title);
-
         return "teacher/course/list";
     }
 
     @GetMapping("/create")
-    public String getCreateCourse(Model model) {
+    public String getCreateCourse(Model model, HttpSession req) {
+        // Check role access site
+        User user = (User) req.getAttribute("userSession");
+        if (user.getRole() != 1)
+            return "AccessDenied";
         model.addAttribute("course", new CourseDTOAdd());
         model.addAttribute("topics", topicService.getTopics());
         return "teacher/course/add";
@@ -87,28 +94,19 @@ public class CourseController {
         if (result.hasErrors()) {
             return "teacher/course/add";
         }
-        // String fileName = ("");
-        // Part filePart = req.getPart("image");
-        // fileName = filePart.getSubmittedFileName();
-        // Resource resource = new ClassPathResource("static/image/");
-        // String absolutePath = resource.getFile().getAbsolutePath();
-        // String filePath = absolutePath + "\\" + fileName;
-        // Path path = Paths.get(filePath);
-        // InputStream inputStream = filePart.getInputStream();
-        // Files.copy(inputStream, path,
-        // StandardCopyOption.REPLACE_EXISTING);
-        // SWP391_ONVID\onlinelearning\target\classes\static\image
-        // SWP391_ONVID\onlinelearning\src\main\resources\static\image
         String projectPath = System.getProperty("user.dir");
         String fileName = "";
         Part filePart = req.getPart("image");
         fileName = filePart.getSubmittedFileName();
-        String relativePath = "\\onlinelearning\\src\\main\\resources\\static\\image";
+        String relativePath = "\\src\\main\\resources\\static\\image";
 
         String filePath = projectPath + relativePath + File.separator + fileName;
         Path path = Paths.get(filePath);
-        InputStream inputStream = filePart.getInputStream();
-        Files.copy(inputStream, path, StandardCopyOption.REPLACE_EXISTING);
+
+        try (InputStream inputStream = filePart.getInputStream()) {
+            Files.copy(inputStream, path, StandardCopyOption.REPLACE_EXISTING);
+        } catch (IOException e) {
+        }
 
         courseDTOAdd.setImageLink(fileName);
         int topic_id = Integer.parseInt(req.getParameter("topic"));
@@ -118,8 +116,21 @@ public class CourseController {
     }
 
     @GetMapping("/edit/{id}")
-    public String getUpdateCourse(Model model, @PathVariable @NotNull Integer id) {
+    public String getUpdateCourse(Model model, @PathVariable @NotNull Integer id, HttpSession req) {
+        // Check role access site
+        User user = (User) req.getAttribute("userSession");
+        if (user.getRole() != 1)
+            return "AccessDenied";
+        // check course exit
         Course course = courseService.getCourseById(id);
+        if (course == null)
+            return "404";
+        // check owner course
+        if (courseService.checkCourseOwner(id))
+            return "AccessDenied";
+        // check published course
+        if (course.getStatus() == 3)
+            return "AccessDenied";
         CourseDTOEdit courseDTEdit = mapper.courseToCourseDtoEdit(course);
         model.addAttribute("course", courseDTEdit);
         model.addAttribute("topics", topicService.getTopics());
@@ -152,13 +163,27 @@ public class CourseController {
     }
 
     @GetMapping("/delete/{id}")
-    public String getDeleteCourse(@PathVariable @NotNull Integer id) {
+    public String getDeleteCourse(@PathVariable @NotNull Integer id, HttpSession req) {
+        // Check role access site
+        User user = (User) req.getAttribute("userSession");
+        if (user.getRole() != 1)
+            return "AccessDenied";
+        // check course exit
+        Course course = courseService.getCourseById(id);
+        if (course == null)
+            return "404";
+        // check owner course
+        if (courseService.checkCourseOwner(id))
+            return "AccessDenied";
+        // check published course
+        if (course.getStatus() == 3)
+            return "AccessDenied";
         courseService.deleteCourse((int) id);
         return "redirect:../list";
     }
 
     @GetMapping("/detail/{id}")
-    public String getDetail(@PathVariable @NotNull Integer id, Model model) {
+    public String getDetail(@PathVariable @NotNull Integer id, Model model, HttpSession req) {
         Course course = courseService.getCourseById(id);
         model.addAttribute("course", course);
         return "teacher/course/detail";
